chore(service-mesh): split service-mesh module
