--- packages/kuma-gui/src/app/service-mesh/README.md
@@ -1,5 +1,0 @@
----
-section: Modules
----
-# service-mesh
-
